Fix wrong usage of std::string::compare while creating URL for ifmap client
std::string::compare function behaves different if strings to be compared are not equal.
For this case we need to define position of element in original string to start comparison
and number of symbols to be compared(common prefix length for both strings).
This change fixes original issue when result is always not equal to 0.

Change-Id: I4162252944fef279e2fa27d7cbd28c9b4a8561a0
Partial-bug: PROD-21151