Merging changes from 5.x to 3.1
Node manager fixes to work with systemd
Changes to make node manager work with both systemd
and supervisord. Abstracted the process information
manager portion of node manager so that it can work
with both systems. Refactor EventManager so that the
node type specific code is abstracted out. Change
supervisord event TICK_60 to a periodic instead so
that common code can be used between supervisord and
systemd.

The systemd process info manager code uses pydbus
which will be installed as part of another commit.

Partial-Bug: #1651777

Change-Id: Iad3afb2ecc0340ff8ec5eb19fd64590f088e6153