Rework attribute check for health monitor objects
According to python documentation hasattr return True
if the string is the name of one of the object’s attributes.
It will be always false because 'hm' object is a dict in our case.
To fix this we need to use 'provider' key of 'hm' dict.

Change-Id: I4e825f507d9ffcf2ebebc5670cc0455ccbfbed5e
Related-Prod: PROD-27478
(cherry picked from commit 75b61f91ae16d255caee2bb697a456d127c22478)