Fix ordering of ST objects reinit to avoid downtime during schema failover
The issue appeared to be in the function “update_static_routes” of RoutingInstanceST class.
It tries to fetch objects from RouteTableST, VirtualMachineInterfaceST and InstanceIpST,
in order to confirm the route target and the route to external network.
But those classes are not yet re-initialized at that moment, hence route target can’t be confirmed and is removed from routing instance.

The solution is to call reinit method of appropriate classes prior it comes to RoutingInstanceST.

“RouteTableST.reinit()” is safe to move up, as all its dependencies are initialized earlier: virtual_networks, logical_routers, service_instances.
InstanceIpST - also safe, as it has refs to virtual_machine_interfaces only.
VirtualMachineInterfaceST - it’s a bit tricky, as it has refs to RoutingInstanceST objects.
Implemented as a workaround currently: just call VirtualMachineInterfaceST.reinit() twice:
one time before RoutingInstanceST reinit, and second time - right after.

PROD-27600 (PROD:27600)

Change-Id: I8d2ef7e65b5fe0acfdab9ce900600376fb20b856