Shrink FreeList container in case of overflow
In high-load scenarios with a huge flow rate the FreeList container may grow
rapidly, using too much RAM.
We actually do not need to store a lot of dummy flow entries in this list, so
it is safe to just shrink it whenever it reaches some threshold, kMaxThreshold.

Here we implement a new method FlowEntryFreeList::Shrink() that will be
triggered when the table FreeList size is exceeding kMaxThreshold (100 * 1000).
We reduce the container size by removing kShrinkSize (4 * 1000) elements, and
we also free the memory allocated for those entries.
Also, removed unused variable grow_count_.

Change-Id: Icac0e8424510a9d3db4f058b527e241eee04fc78
Related-Prod: PROD-31653