New config option for Contrail vRouter Agent: disable_stats_collection
Here we introduce new Boolean option:
[DEFAULT]
disable_stats_collection=true

This option controlls - do we want Agent to collect and send to Analytics
statistics related to VMIs and VNs.
It is turned off by default, to preserve original behavior as is.

Motivation for new this option:
Consider a cluster with hundreds vRouter nodes.
And each node has a lot of workloads on it: dozens of instances, VRFs,
hundreds of VMIs.
With such scenario, Analytics is overloaded by huge amount of VMI/VN stats
coming from all vRouter nodes.

Related-PROD: PROD-29404
Change-Id: I6ee59cc8b8dd0bedd9d7e39c38779b936aaee60b